Remove unused big uint utilities from plugin sdk
